+++ benchmark/citation/sgc.py
@@ -5,6 +5,7 @@ import torch.nn.functional as F
 from citation import get_planetoid_dataset, random_planetoid_splits, run
 
 from torch_geometric.nn import SGConv
+from torch_geometric.profile import rename_profile_file
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--dataset', type=str, required=True)
@@ -16,6 +17,8 @@ parser.add_argument('--weight_decay', type=float, default=0.0005)
 parser.add_argument('--early_stopping', type=int, default=10)
 parser.add_argument('--no_normalize_features', action='store_true')
 parser.add_argument('--K', type=int, default=2)
+parser.add_argument('--inference', action='store_true')
+parser.add_argument('--profile', action='store_true')
 args = parser.parse_args()
 
 
@@ -37,4 +40,8 @@ class Net(torch.nn.Module):
 dataset = get_planetoid_dataset(args.dataset, not args.no_normalize_features)
 permute_masks = random_planetoid_splits if args.random_splits else None
 run(dataset, Net(dataset), args.runs, args.epochs, args.lr, args.weight_decay,
-    args.early_stopping, permute_masks)
+    args.early_stopping, args.inference, args.profile, permute_masks)
+
+if args.profile:
+    rename_profile_file('citation', SGConv.__name__, args.dataset,
+                        str(args.random_splits))