+++ benchmark/points/point_cnn.py
@@ -7,6 +7,7 @@ from points.train_eval import run
 from torch.nn import Linear as Lin
 
 from torch_geometric.nn import XConv, fps, global_mean_pool
+from torch_geometric.profile import rename_profile_file
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--epochs', type=int, default=200)
@@ -15,6 +16,8 @@ parser.add_argument('--lr', type=float, default=0.001)
 parser.add_argument('--lr_decay_factor', type=float, default=0.5)
 parser.add_argument('--lr_decay_step_size', type=int, default=50)
 parser.add_argument('--weight_decay', type=float, default=0)
+parser.add_argument('--inference', action='store_true')
+parser.add_argument('--profile', action='store_true')
 args = parser.parse_args()
 
 
@@ -60,4 +63,8 @@ class Net(torch.nn.Module):
 train_dataset, test_dataset = get_dataset(num_points=1024)
 model = Net(train_dataset.num_classes)
 run(train_dataset, test_dataset, model, args.epochs, args.batch_size, args.lr,
-    args.lr_decay_factor, args.lr_decay_step_size, args.weight_decay)
+    args.lr_decay_factor, args.lr_decay_step_size, args.weight_decay,
+    args.inference, args.profile)
+
+if args.profile:
+    rename_profile_file('points', XConv.__name__)