+++ benchmark/points/train_eval.py
@@ -3,15 +3,16 @@ import time
 import torch
 import torch.nn.functional as F
 from torch.optim import Adam
+from torch.profiler import ProfilerActivity, profile
 
 from torch_geometric.loader import DataLoader
+from torch_geometric.profile import trace_handler
 
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
 
-def run(train_dataset, test_dataset, model, epochs, batch_size, lr,
-        lr_decay_factor, lr_decay_step_size, weight_decay):
-
+def run_train(train_dataset, test_dataset, model, epochs, batch_size, lr,
+              lr_decay_factor, lr_decay_step_size, weight_decay):
     model = model.to(device)
     optimizer = Adam(model.parameters(), lr=lr, weight_decay=weight_decay)
 
@@ -40,6 +41,43 @@ def run(train_dataset, test_dataset, model, epochs, batch_size, lr,
                 param_group['lr'] = lr_decay_factor * param_group['lr']
 
 
+@torch.no_grad()
+def run_inference(test_dataset, model, epochs, batch_size, profiling):
+    model = model.to(device)
+    test_loader = DataLoader(test_dataset, batch_size, shuffle=False)
+
+    for epoch in range(1, epochs + 1):
+        if epoch == epochs:
+            if torch.cuda.is_available():
+                torch.cuda.synchronize()
+            t_start = time.time()
+
+        inference(model, test_loader, device)
+
+        if epoch == epochs:
+            if torch.cuda.is_available():
+                torch.cuda.synchronize()
+            t_end = time.time()
+            duration = t_end - t_start
+            print(f'End-to-End Inference Time: {duration:.8f}s', flush=True)
+
+    if profiling:
+        with profile(activities=[ProfilerActivity.CPU, ProfilerActivity.CUDA],
+                     on_trace_ready=trace_handler) as p:
+            inference(model, test_loader, device)
+            p.step()
+
+
+def run(train_dataset, test_dataset, model, epochs, batch_size, lr,
+        lr_decay_factor, lr_decay_step_size, weight_decay, inference,
+        profiling):
+    if not inference:
+        run_train(train_dataset, test_dataset, model, epochs, batch_size, lr,
+                  lr_decay_factor, lr_decay_step_size, weight_decay)
+    else:
+        run_inference(test_dataset, model, epochs, batch_size, profiling)
+
+
 def train(model, optimizer, train_loader, device):
     model.train()
 
@@ -52,6 +90,7 @@ def train(model, optimizer, train_loader, device):
         optimizer.step()
 
 
+@torch.no_grad()
 def test(model, test_loader, device):
     model.eval()
 
@@ -63,3 +102,11 @@ def test(model, test_loader, device):
     test_acc = correct / len(test_loader.dataset)
 
     return test_acc
+
+
+@torch.no_grad()
+def inference(model, test_loader, device):
+    model.eval()
+    for data in test_loader:
+        data = data.to(device)
+        model(data.pos, data.batch)