+++ torch_geometric/profile/__init__.py
@@ -1,4 +1,5 @@
 from .profile import profileit, timeit, get_stats_summary
+from .profile import trace_handler, rename_profile_file
 from .utils import count_parameters
 from .utils import get_model_size
 from .utils import get_data_size
@@ -10,6 +11,8 @@ __all__ = [
     'profileit',
     'timeit',
     'get_stats_summary',
+    'trace_handler',
+    'rename_profile_file',
     'count_parameters',
     'get_model_size',
     'get_data_size',